+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 
 ### Next release
 
+* 🔨 Trakt: links to more information and ratings go to their new web app.
 * 🔧 Movies: cache genres and trailer to database, display cached data immediately before potentially slow network requests.
 
 ## Version 2026.3